Laying the foundation for adding labels.
