+++ api/doc/README.md
@@ -4,3 +4,4 @@
 * [client commands](./client.md)
 * [golang tips](./tips.md)
 * [Try curl requests](./play.along.md)
+* [Manual testing of endpoints](./manual.test.md)
